refactor third round
